refactor: env resolution to separate file
