+++ src/lib/auth.ts
@@ -1,7 +1,7 @@
 import { db } from '../db';
 import { createAuth } from './auth-config';
 
-export { createAuth, getAuthEnv } from './auth-config';
+export { createAuth } from './auth-config';
 
 export const auth = createAuth({
   database: db,
